requirements.txt
google-adk==1.31.1
google-genai==1.73.1
google-cloud-aiplatform==1.148.1
mcp==1.27.0
python-dotenv==1.2.2
pydantic==2.13.3
gspread==6.1.4
google-api-python-client>=2.120.0
